apps/web/src/features/application-shell/hooks/use-application-shell.ts
interface UseApplicationShellOptions {
  /** Optional configuration for the application-shell hook */
  config?: {
    /** Enable/disable feature */
    enabled?: boolean
    /** Timeout in milliseconds */
    timeout?: number
  }
}

interface UseApplicationShellResult {
  /** Loading state */
  isLoading: boolean
  /** Error state */
  error: Error | null
  /** Reset the hook state */
  reset: () => void
}

export const useApplicationShell = ({
  config,
}: UseApplicationShellOptions = {}): UseApplicationShellResult => {
  return {
    isLoading: false,
    error: null,
    reset: () => {
      // Reset implementation
    },
  }
}
