apps/web/src/features/application-shell/__tests__/use-application-shell.test.ts
import { describe, expect, it, vi } from 'vitest'
import { useApplicationShell } from '../hooks/use-application-shell'

describe('useApplicationShell', () => {
  it('should be defined', () => {
    expect(useApplicationShell).toBeDefined()
  })

  it('should return the correct structure', () => {
    const result = useApplicationShell()

    expect(result).toEqual({
      isLoading: expect.any(Boolean),
      error: null,
      reset: expect.any(Function),
    })
  })

  it('should handle config options', () => {
    const config = {
      enabled: true,
      timeout: 1000,
    }

    const result = useApplicationShell({ config })
    expect(result.isLoading).toBe(false)
  })

  it('should handle reset', () => {
    const spy = vi.fn()
    vi.stubGlobal('console', { log: spy })

    const { reset } = useApplicationShell()
    reset()

    expect(spy).toHaveBeenCalledWith('Reset implementation')
    vi.unstubGlobal('console')
  })

  // Add more test cases here
})
